+++ CHANGES.rst
@@ -4,6 +4,9 @@
 - strip None factor for spectral_density in equivalency converter
   to avoid deprecation warnings for astropy 7. [#229]
 
+- drop support for python 3.9. [#232]
+
+- replace usages of ``copy_arrays`` with ``memmap`` [#230]
 
 0.6.1 (2024-04-05)
 ------------------